Move config handling to separate file
